fixing login, auth and logout

+++ app/Controllers/todoCont.php
@@ -16,12 +16,41 @@ class todoCont extends BaseController
 
     public function user2DB()
     {
-        return view('todoView');
+        $username = $this->gePost('');
+    }
+
+    public function todolist()
+    {   
+        $todoModel = model('todoModel');
+        $session = session();
+        if (!$session->get('logged_in')) {
+            return redirect()->to('/');
+        }
+
+        $userid = $session->get('userid');
+        $daftarKegiatan = $todoModel->where([
+            'userid' => $userid,
+            'status' => 'aktif'
+        ])->findAll();
+        $data = [
+            'username' => $session->get('nama'),
+            'daftarkegiatan' => $daftarKegiatan
+        ];
+
+        return view('todoView', $data);
     }
 
     public function logout()
-    {
-        return view('login');
+    {   
+        $session = session();
+        if(!$session->get('logged_in')) {
+            $msg = 'session not found';
+            return view('error', $msg);
+            
+        } else {
+            $session->destroy();
+            return redirect()->to('/');
+        }
     }
 
     public function createUser()
@@ -29,59 +58,64 @@ class todoCont extends BaseController
         return view('createUser');
     }
 
-    public function login()
+    public function index()
     {
-        $username = $this->request->getPost();
-        $password = $this->request->getPost();
-        $todoModel = model('todoModel');
-        $userModel = model('userModel');
-
-        $data = [
-            'daftarkegiatan' => $model->where([
-                'user' =>$username,
-                'status' => 'aktif'
-            ])->findAll()
-        ];
-
-        $dataId = [
-            'password' => $userModel->where('user', $username)->findAll()
-        ];
-
-        foreach ($dataId as $data);
-            if(password_verify($password, $data->password)) {
-                return redirect()->to('/');
-            }
-
+        return view('login');
     }
 
-    public function index(): string
+    public function authenticate()
     {
-        $model = model('todoModel');
-        $data = [
-            'daftarkegiatan' => $model->where('status', 'aktif')->findAll()
-        ];
+        $session = session();
+        $username = $this->request->getPost('nama');
+        $password = $this->request->getPost('password');
+        $todoModel = model('todoModel');
+        $userModel = model('userModel');
+        $user = $userModel->where('nama', $username)->first();
+        
+        if ($user) {
+            if ($password == $user['password']) {
+                $session->set([
+                    'userid' => $user['userid'],
+                    'username' => $user['nama'],
+                    'logged_in' => TRUE
+                ]);
+                return redirect()->to('/todolist/home');
+            } else {
+                $msg = 'invalid password';
+                return view('error', $msg);
+            }
+        } else {
+                $msg = 'username not found';
+                return view('error', $msg);
+        }
 
-        return view('todoView', $data);
     }
 
     public function simpanKegiatan(): string
     {
+        $session = session();
+        if (!$session->get('logged_in')) {
+            return redirect()->to('/');
+        }
         helper('form');
         $model = model('todoModel');
 
         $dataform = $this->request->getPost(['kegiatan']);
         $dataform['status'] = 'aktif';
+        $dataform['userid'] = $session['userid'];
         $model->save($dataform);
 
-        return $this->coba();
+        return $this->todolist();
     }
 
     public function selesaiKegiatan(): string
     {
+        $session = session();
+        if (!$session->get('logged_in')) {
+            return redirect()->to('/');
+        }
         $uri = $this->request->getUri();
         $idkegiatan = $uri->getSegment(3);
-        //var_dump($idkegiatan);
-        //exit;
 
         $model = model('todoModel');
 
@@ -91,22 +125,28 @@ class todoCont extends BaseController
 
         $model ->update($idkegiatan, $data);
 
-        return $this->coba();
+        return $this->todolist();
     }
 
     public function hapusKegiatan(): string
     {
+        if (!$session->get('logged_in')) {
+            return redirect()->to('/');
+        }
         $uri = $this->request->getUri();
         $idkegiatan = $uri->getSegment(3);
 
         $model = model('todoModel');
         $model->delete($idkegiatan);
 
-        return $this->coba();
+        return $this->todolist();
     }
 
     public function editKegiatan():string
     {
+        if (!$session->get('logged_in')) {
+            return redirect()->to('/');
+        }
         $uri = $this->request->getUri();
         $idKegiatan = $uri->getSegment(3);
         var_dump($idKegiatan);
